+++ features/search/search-by-conversation-membership/search-by-conversation-membership.test.ts
@@ -1,0 +1,11 @@
+import { searchByConversationMembership } from "./search-by-conversation-membership";
+
+describe("searchByConversationMembership", () => {
+  test("searchByConversationMembership", async () => {
+    // expect(false).toBe(true);
+    const result = await searchByConversationMembership({
+      searchQuery: "test",
+    });
+    expect(result).toBeDefined();
+  });
+});
