+++ features/search/search-by-conversation-membership/search-by-conversation-membership.ts
@@ -1,191 +1,15 @@
-import { queryOptions, useQuery } from "@tanstack/react-query";
-
-import { IProfileSocials } from "@/features/profiles/profile-types";
-import {
-  searchByConversationMembershipQueryKey,
-  userSearchQueryKey,
-} from "./QueryKeys";
-
-type ISearchArgs = {
-  searchQuery: string;
-  addressesToOmit: string[];
-};
-
 import { getCurrentAccount } from "@/data/store/accountsStore";
-import { getAddressForPeer, isSupportedPeer } from "@/utils/evm/address";
-import { getCleanAddress } from "@/utils/evm/getCleanAddress";
-import { searchProfilesForCurrentAccount } from "@/utils/api/profiles";
-import {
-  getProfileSocialsQueryData,
-  setProfileRecordSocialsQueryData,
-} from "@/queries/useProfileSocialsQuery";
-import { isEmptyObject } from "@/utils/objects";
-import { shortAddress } from "@/utils/strings/shortAddress";
-import { accountCanMessagePeer } from "@/features/consent/account-can-message-peer";
+import { getProfileSocialsQueryData } from "@/queries/useProfileSocialsQuery";
 import logger from "@/utils/logger";
-import { getConversationsQueryData } from "./use-conversations-query";
-import { ensureGroupMembersQueryData } from "./useGroupMembersQuery";
+import { getConversationsQueryData } from "../../../queries/use-conversations-query";
+import { ensureGroupMembersQueryData } from "../../../queries/useGroupMembersQuery";
 import { isConversationGroup } from "@/features/conversation/utils/is-conversation-group";
 import { isConversationDm } from "@/features/conversation/utils/is-conversation-dm";
 import { ConversationVersion } from "@xmtp/react-native-sdk";
-import { getPreferredName } from "@/utils/profile";
 import { getReadableProfile } from "@/utils/getReadableProfile";
+import { IConversationMembershipSearchResult } from "@/features/search/search.types";
 
-/**
- * Handles searching when searchQuery is a supported domain/ENS
- */
-async function handlePeerSearch(
-  searchQuery: string
-): Promise<IConvosUsersSearchResult> {
-  logger.info(`[Search] Starting peer search for query: ${searchQuery}`);
-  logger.info(`[Search] Query ${searchQuery} is a supported peer format`);
-  const resolvedAddress = await getAddressForPeer(searchQuery);
-
-  if (!resolvedAddress) {
-    logger.info(`[Search] No address resolved for ${searchQuery}`);
-    return {
-      convosSearchResults: {},
-      message: "No address has been set for this domain.",
-    };
-  }
-
-  const address = getCleanAddress(resolvedAddress);
-  logger.info(`[Search] Resolved address ${address} for ${searchQuery}`);
-  logger.info(`[Search] Checking if ${address} is on XMTP`);
-
-  const addressIsOnXmtp = await accountCanMessagePeer({
-    account: getCurrentAccount()!,
-    peer: address,
-  });
-
-  if (!addressIsOnXmtp) {
-    logger.info(`[Search] ${address} is not on XMTP`);
-    return {
-      message: `${shortAddress(searchQuery)} is not on the XMTP network yet`,
-      convosSearchResults: {},
-    };
-  }
-
-  logger.info(`[Search] ${address} is on XMTP, fetching profiles`);
-  const profiles = await searchProfilesForCurrentAccount(address);
-  logger.info(`[Search] Profiles fetched:`, JSON.stringify(profiles, null, 2));
-
-  if (!isEmptyObject(profiles)) {
-    logger.info(`[Search] Found profiles for ${address}, setting profile data`);
-    setProfileRecordSocialsQueryData(profiles);
-    return {
-      message: "",
-      convosSearchResults: profiles,
-    };
-  }
-
-  logger.info(
-    `[Search] No profiles found for ${address}, returning just address`
-  );
-  const justAddress: Record<string, IProfileSocials> = {
-    [address]: { address },
-  };
-  return {
-    message: "address is on xmtp but not on converse yet",
-    convosSearchResults: justAddress,
-  };
-}
-
-/**
- * Handles searching when searchQuery is a regular string (not a peer)
- */
-async function handleGeneralSearch(searchQuery: string) {
-  logger.info(`[Search] Starting general search for query: ${searchQuery}`);
-  logger.info(`[Search] Searching profiles for query: ${searchQuery}`);
-  const profiles = await searchProfilesForCurrentAccount(searchQuery);
-  logger.info(`[Search] Profiles found:`, JSON.stringify(profiles, null, 2));
-
-  if (!isEmptyObject(profiles)) {
-    logger.info(`[Search] Found profiles for query ${searchQuery}`);
-    setProfileRecordSocialsQueryData(profiles);
-    const filteredProfiles = { ...profiles };
-    logger.info(
-      `[Search] Filtered profiles:`,
-      JSON.stringify(filteredProfiles, null, 2)
-    );
-    return {
-      message: "",
-      convosSearchResults: filteredProfiles,
-    };
-  }
-
-  logger.info(`[Search] No profiles found for query ${searchQuery}`);
-  return {
-    message: `They're not here\nInvite them?`,
-    convosSearchResults: {},
-  };
-}
-
-/**
- * Users our backend to search for users in the Convo network
- *
- * @see we also have methods for searching for users by current conversation cached locally
- */
-export async function searchConvosUsers({
-  searchQuery,
-}: {
-  searchQuery: string;
-}): Promise<IConvosUsersSearchResult> {
-  logger.info(`[Search] Starting search for query: ${searchQuery}`);
-
-  if (searchQuery.length === 0) {
-    logger.info(`[Search] Empty search query, returning empty results`);
-    return {
-      message: "",
-      convosSearchResults: {},
-    };
-  }
-
-  if (isSupportedPeer(searchQuery)) {
-    return handlePeerSearch(searchQuery);
-  }
-
-  return handleGeneralSearch(searchQuery);
-}
-
-const SearchResultStaleTime = 1000 * 10; // 10 seconds
-
-type GroupNameSearchResult = {
-  groupName: string;
-  groupId: string;
-  groupImageUri: string;
-  firstThreeMemberNames: string[];
-};
-
-type GroupMemberNameSearchResult = GroupNameSearchResult & {
-  memberNameFromGroup: string;
-};
-
-export type IUserSearchResults = IConvosUsersSearchResult &
-  IConversationMembershipSearchResult;
-
-export type IConvosUsersSearchResult = {
-  convosSearchResults?: Record<string, IProfileSocials>;
-  message?: string;
-};
-
-export function getConvosUsersSearchQueryOptions(searchQuery: string) {
-  logger.info(`[Search] Creating query options for search: ${searchQuery}`);
-  return queryOptions({
-    queryKey: userSearchQueryKey(searchQuery),
-    queryFn: () => searchConvosUsers({ searchQuery }),
-    enabled: !!searchQuery,
-    staleTime: SearchResultStaleTime,
-  });
-}
-
-export type IConversationMembershipSearchResult = {
-  existingDmSearchResults?: Record<string, IProfileSocials>;
-  existingGroupMemberNameSearchResults?: GroupMemberNameSearchResult[];
-  existingGroupNameSearchResults?: GroupNameSearchResult[];
-};
-
-async function searchByConversationMembership({
+export async function searchByConversationMembership({
   searchQuery,
 }: {
   searchQuery: string;
@@ -410,88 +234,3 @@ async function searchByConversationMembership({
   );
   return result;
 }
-
-export function getSearchByConversationMembershipQueryOptions(
-  searchQuery: string
-) {
-  logger.info(
-    `[Search] Creating conversation membership query options for: ${searchQuery}`
-  );
-  return queryOptions({
-    queryKey: searchByConversationMembershipQueryKey(searchQuery),
-    queryFn: () => searchByConversationMembership({ searchQuery }),
-    enabled: !!searchQuery,
-    // staleTime: SearchResultStaleTime,
-    staleTime: 0,
-  });
-}
-
-export const useSearchUsersQuery = (args: ISearchArgs) => {
-  const { searchQuery } = args;
-  const currentAccount = getCurrentAccount()!;
-  const currentAccountAddress = getCleanAddress(currentAccount);
-  const allAddressesToOmit = [
-    ...args.addressesToOmit,
-    currentAccountAddress,
-  ].map((address) => address.toLowerCase());
-
-  logger.info(
-    `[Search] Starting search query with args:`,
-    JSON.stringify(args, null, 2)
-  );
-  logger.info(
-    `[Search] Addresses to omit:`,
-    JSON.stringify(allAddressesToOmit, null, 2)
-  );
-
-  const {
-    data: { convosSearchResults, message } = {},
-    isLoading: areSearchResultsLoading,
-  } = useQuery({
-    ...getConvosUsersSearchQueryOptions(searchQuery),
-    select: (data) => {
-      const filteredResults = { ...data };
-      if (filteredResults.convosSearchResults) {
-        Object.keys(filteredResults.convosSearchResults).forEach((address) => {
-          if (allAddressesToOmit.includes(address.toLowerCase())) {
-            logger.info(`[Search] Omitting address from results: ${address}`);
-            delete filteredResults.convosSearchResults?.[address];
-          }
-        });
-      }
-
-      return filteredResults;
-    },
-  });
-
-  const {
-    data: {
-      existingDmSearchResults,
-      existingGroupMemberNameSearchResults,
-      existingGroupNameSearchResults,
-    } = {},
-  } = useQuery(getSearchByConversationMembershipQueryOptions(searchQuery));
-
-  const searchResults: IUserSearchResults = {
-    existingDmSearchResults,
-    existingGroupMemberNameSearchResults,
-    existingGroupNameSearchResults,
-    convosSearchResults,
-    message,
-  };
-
-  logger.info(
-    `[Search] Final combined search results:`,
-    JSON.stringify(
-      { ...searchResults, convosSearchResults: undefined },
-      null,
-      2
-    )
-  );
-
-  return {
-    searchResults,
-    hasSearchResults: true,
-    areSearchResultsLoading,
-  };
-};
